Refactor handler API and modernize bindings
Rename handler types to BoxedAction/BoxedActionOnce/SharedAction
Introduce new helper constructors and shared-action wrapper; update FFI
Replace Binding::container with Binding::f64/f32/bool
Simplify numeric suffixes (2.0_f32 -> 2.0) and adjust imports/usages
Remove audio-visualizer WGSL shader and add conditional rendering docs

+++ examples/hover/src/lib.rs
@@ -156,7 +156,7 @@ fn reactive_cursor_section(is_dragging: &Binding<bool>) -> impl View {
             )
             .opacity(
                 is_dragging_opacity
-                    .select(0.8 as f32, 1.0)
+                    .select(0.8, 1.0)
                     .with(Animation::default()),
             )
             .with_state(is_dragging)
